Reduce comparisons for function Eq instances
In tests we have `Eq` instances for function types so that we can test
laws that require `Eq` instances. They work by generating a bunch of
random values and checking that the two functions return equivalent
results for those values. This results in essentially nested loops of
generating values: once in the Scalacheck `forAll` for the property
being tested, and another for generating the random values to check the
equivalence of the functions.

This nested loop of generation makes tests slower, and Cats builds are
often timing out. Since these checks are run on every CI build, and are
already the inner part of a generation loop in tests, I think that it's
probably safe to turn down the dial on the number of values generated in
the interest of stabilizing the build.